Watch playdar pid for termination status
So the enable button updates if the playdar process terminates even if we
didn't start or stop it ourselves.

The one instance this still doesn't work for though is if the playdar process
was started outside of the prefpane. Eg. from a terminal. I can't find a way
to determine this without polling at this point. But I might just implement
that.